# improve VS update performance after project changes 