--- a/vsintegration/src/FSharp.Editor/Options/EditorOptions.fs
+++ b/vsintegration/src/FSharp.Editor/Options/EditorOptions.fs
@@ -10,8 +10,8 @@ open OptionsUIHelpers
 
 module DefaultTuning = 
     let SemanticColorizationInitialDelay = 0 (* milliseconds *)
-    let UnusedDeclarationsAnalyzerInitialDelay = 1000 (* milliseconds *)
-    let UnusedOpensAnalyzerInitialDelay = 2000 (* milliseconds *)
+    let UnusedDeclarationsAnalyzerInitialDelay = 0 (* 1000 *) (* milliseconds *)
+    let UnusedOpensAnalyzerInitialDelay = 0 (* 2000 *) (* milliseconds *)
     let SimplifyNameInitialDelay = 2000 (* milliseconds *)
     let SimplifyNameEachItemDelay = 5 (* milliseconds *)