This is a set of improvements to VS perf, initially for https://github.com/Microsoft/visualfsharp/issues/3094 but also https://github.com/Microsoft/visualfsharp/issues/2107 and https://github.com/Microsoft/visualfsharp/issues/3054, and indeed any situation where you're editing code across multiple files in a project.

1. Remove the 1 second pauses before background work is started when the queue is empty, and use a 5msec pause instead.  The intention of the 1sec pause was to allow foreground UI intellisense requests to come in before starting background work.  However we are relying on background work to happen to propagate some things in the UI  (e.g. to push through changes in solution options by incrementally type checking the project and triggering a Reanalyze).  Further, I suspect the 1 second pauses were happening all the time, especially when SimplifyName diagnostic analyzer was being used.  

2. Remove the startup pauses for some of the analyzers. Normally these analyzers get run in parallel, and the async startup pause was effectively a cheapshot way of specifying priorities.  However these analyers get run in sequence after a Reanalyze has been requested.  Further they get run before semantic colorization is done. This means no colors get updated until all of them have been completed

3. Disable the SimplifyName analyzer. This can take a a long time to run because the names are checked one-by-one (and there is currently a pause checking each name). I strongly suspect that this is a real problem and a major cause of "lag" in different ways.  If analyzers are run in sequence (e.g. when Reanalyze is called) , then any long-running analyzers are a real problem 

@vasily-kirichenko From what I've seen in debugging we can't afford to have SimplifyName analyzer on if document diagnostic analyzers are being run in sequence in response to Reanalyze.  Perhaps we could have an option to explicitly enable it, we could double check whether Roslyn really is running the analyzers in sequence and why.  I'm a bit concerned about the other analyzers too for large files.

All in all the changes in this PR _seem_ to significantly improve the performance of red-squiggly update speed in the face of 
* edits in other files
* project configuration changes (from 10sec down to 5sec, see https://github.com/Microsoft/visualfsharp/issues/3094)
* project option changes