@dsyme 2017-06-22
This also includes a fix to https://github.com/Microsoft/visualfsharp/issues/2107, see [this line](https://github.com/Microsoft/visualfsharp/pull/3238/files#diff-cdf0a4fbb7615b8029d78133924259c3R451). After this change we get [these results](https://github.com/Microsoft/visualfsharp/issues/2107#issuecomment-310244542) v VS2015