@dsyme 2017-06-22
@dotnet-bot test Windows_NT Release_ci_part2 Build please