@dsyme 2017-06-22
> running analyzers serially: was a deliberate decision for in-proc analyzers, since running them in parallel ended up generating enough garbage that collections significantly impacted UI thread responsiveness while typing.

@Pilchie @heejaechang Could we open a Roslyn issue about this?  I think it's a significant issue. For F# we have more or less been assuming that all these analyzers get run in parallel on more or less every code edit (and get cancelled if another edit occurs).  We put in some "async sleep" operations to effectively change their priority, delaying the low priority ones.  Now that we understand that the document diagnostic analyzers are actually getting run serially in response to "Reanalyze" we will need to reconsider and scale back how much work we do in the analyzers.