@dsyme 2017-06-22
@KevinRansom @cartermp This is ready.  It's an important set of fixes (and one feature removal) that we should include ASAP.