@dsyme 2017-06-22
> Long term, I think we should try to push for an API to open up here so that we don't have to rely on DocumentDiagnosticAnalyzer. I think it's only used in a few small places in Roslyn, and not intended to be run in parallel.

I don't know Roslyn well enough, but I'd somehow be surprised if we can't implement these as some other kind of analyzer