@heejaechang 2017-06-22
we are in talk with data scientist (@roofkitty) to come up with formula to track performance of each analyzer. once we have that in place, we are planning to use that data to order analyzers in real time. cheaper one first, slow one last.

also, we could add some metadata to Document/ProjectDiagnosticAnalyzer to let people to specific ordering such as Low priority.