@dsyme 2017-06-22
> we are in talk with data scientist (@roofkitty) to come up with formula to track performance of each analyzer. once we have that in place, we are planning to use that data to order analyzers in real time. cheaper one first, slow one last.

This approach won't work for F#.  The first one will always be the slow one, since FSharp.Compiler.Service caches the parse/check request.  So you will be blaming the first one, you will move it to the back of the queue, and then you will blame the next one etc. etc.  That is, you are assuming the performance of the analyzers is independent, when it isn't.

> also, we could add some metadata to Document/ProjectDiagnosticAnalyzer to let people to specific ordering such as Low priority.

That will work, thanks