@heejaechang 2017-06-22
@dsyme we are in same situation. so for the problem, we make sure we always run compiler as the very first analyzer so that it can absorb all those lazy-ness cost. and we ignore that analyzer from perf analysis so only other analyzer gets proper ordering.

looks like we also need to have a way to mark [required] or high priority analyzers ..