@dsyme 2017-08-16
Reopening per https://github.com/Microsoft/visualfsharp/issues/3267#issuecomment-322833058 - one line of the fix for this has been reverted - we need to re-check that the solution load performance results are still valid